+++ v2/internal/provider/resource_router_bfd_neighbor.go
@@ -1,0 +1,265 @@
+// Unofficial Fortinet Terraform Provider
+// Generated from templates using FortiOS v6.4.0,v7.0.4 schemas
+// Maintainers:
+// Justin Roberts (@poroping)
+
+// Description: Neighbor.
+
+package provider
+
+import (
+	"context"
+	"log"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
+	"github.com/poroping/forti-sdk-go/v2/models"
+	"github.com/poroping/terraform-provider-fortios/v2/utils"
+)
+
+func resourceRouterBfdNeighbor() *schema.Resource {
+	return &schema.Resource{
+		Description: "Neighbor.",
+
+		CreateContext: resourceRouterBfdNeighborCreate,
+		ReadContext:   resourceRouterBfdNeighborRead,
+		UpdateContext: resourceRouterBfdNeighborUpdate,
+		DeleteContext: resourceRouterBfdNeighborDelete,
+
+		Importer: &schema.ResourceImporter{
+			StateContext: schema.ImportStatePassthroughContext,
+		},
+
+		Schema: map[string]*schema.Schema{
+			"vdomparam": {
+				Type:        schema.TypeString,
+				Description: "Specifies the vdom to which the resource will be applied when the FortiGate unit is running in VDOM mode. If you want to inherit the VDOM configuration of the provider, do not set this parameter.",
+				Optional:    true,
+				ForceNew:    true,
+			},
+			"allow_append": {
+				Type:        schema.TypeBool,
+				Description: "If set the provider will overwrite existing resources with the same mkey instead of erroring. Useful for brownfield implementations. Use with caution!",
+				Optional:    true,
+				Default:     false,
+			},
+			"interface": {
+				Type:         schema.TypeString,
+				ValidateFunc: validation.StringLenBetween(0, 15),
+
+				Description: "Interface name.",
+				Optional:    true,
+				Computed:    true,
+			},
+			"ip": {
+				Type:         schema.TypeString,
+				ValidateFunc: validation.IsIPv4Address,
+
+				Description: "IPv4 address of the BFD neighbor.",
+				ForceNew:    true,
+				Required:    true,
+			},
+		},
+	}
+}
+
+func resourceRouterBfdNeighborCreate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	c := meta.(*apiClient).Client
+	var diags diag.Diagnostics
+	var err error
+	// c.Retries = 1
+	urlparams := &models.CmdbRequestParams{}
+	vdomparam := ""
+	if v, ok := d.GetOk("vdomparam"); ok {
+		if s, ok := v.(string); ok {
+			vdomparam = s
+		}
+	}
+	urlparams.Vdom = vdomparam
+
+	allow_append := false
+	if v, ok := d.GetOk("allow_append"); ok {
+		if b, ok := v.(bool); ok {
+			allow_append = b
+		}
+	}
+	urlparams.AllowAppend = &allow_append
+
+	mkey := ""
+	key := "ip"
+	if v, ok := d.GetOk(key); ok {
+		mkey = utils.ParseMkey(v)
+		if mkey == "" && allow_append {
+			return diag.Errorf("error creating RouterBfdNeighbor resource: %q must be set if \"allow_append\" is true", key)
+		}
+	}
+
+	obj, diags := getObjectRouterBfdNeighbor(d, c.Config.Fv)
+	if diags.HasError() {
+		return diags
+	}
+
+	o, err := c.Cmdb.CreateRouterBfdNeighbor(obj, urlparams)
+
+	if err != nil {
+		e := diag.FromErr(err)
+		return append(diags, e...)
+	}
+
+	if o.Mkey != nil {
+		d.SetId(utils.ParseMkey(o.Mkey))
+	} else {
+		d.SetId("RouterBfdNeighbor")
+	}
+
+	return resourceRouterBfdNeighborRead(ctx, d, meta)
+}
+
+func resourceRouterBfdNeighborUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	mkey := d.Id()
+	c := meta.(*apiClient).Client
+	// c.Retries = 1
+	urlparams := &models.CmdbRequestParams{}
+	vdomparam := ""
+	if v, ok := d.GetOk("vdomparam"); ok {
+		if s, ok := v.(string); ok {
+			vdomparam = s
+		}
+	}
+	urlparams.Vdom = vdomparam
+
+	obj, diags := getObjectRouterBfdNeighbor(d, c.Config.Fv)
+	if diags.HasError() {
+		return diags
+	}
+
+	o, err := c.Cmdb.UpdateRouterBfdNeighbor(mkey, obj, urlparams)
+	if err != nil {
+		return diag.Errorf("error updating RouterBfdNeighbor resource: %v", err)
+	}
+
+	// log.Printf(strconv.Itoa(c.Retries))
+	if o.Mkey != nil {
+		d.SetId(utils.ParseMkey(o.Mkey))
+	} else {
+		d.SetId("RouterBfdNeighbor")
+	}
+
+	return resourceRouterBfdNeighborRead(ctx, d, meta)
+}
+
+func resourceRouterBfdNeighborDelete(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	mkey := d.Id()
+
+	c := meta.(*apiClient).Client
+	// c.Retries = 1
+
+	urlparams := &models.CmdbRequestParams{}
+	vdomparam := ""
+	if v, ok := d.GetOk("vdomparam"); ok {
+		if s, ok := v.(string); ok {
+			vdomparam = s
+		}
+	}
+	urlparams.Vdom = vdomparam
+
+	err := c.Cmdb.DeleteRouterBfdNeighbor(mkey, urlparams)
+	if err != nil {
+		return diag.Errorf("error deleting RouterBfdNeighbor resource: %v", err)
+	}
+
+	d.SetId("")
+
+	return nil
+}
+
+func resourceRouterBfdNeighborRead(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
+	mkey := d.Id()
+
+	c := meta.(*apiClient).Client
+	// c.Retries = 1
+
+	urlparams := &models.CmdbRequestParams{}
+	vdomparam := ""
+	if v, ok := d.GetOk("vdomparam"); ok {
+		if s, ok := v.(string); ok {
+			vdomparam = s
+		}
+	}
+	urlparams.Vdom = vdomparam
+
+	ptp := true
+	urlparams.PlainTextPassword = &ptp
+
+	o, err := c.Cmdb.ReadRouterBfdNeighbor(mkey, urlparams)
+	if err != nil {
+		return diag.Errorf("error reading RouterBfdNeighbor resource: %v", err)
+	}
+
+	if o == nil {
+		log.Printf("[WARN] resource (%s) not found, removing from state", d.Id())
+		d.SetId("")
+		return nil
+	}
+
+	sort := false
+	if v, ok := d.GetOk("dynamic_sort_subtable"); ok {
+		if b, ok := v.(bool); ok {
+			sort = b
+		}
+	}
+
+	diags := refreshObjectRouterBfdNeighbor(d, o, c.Config.Fv, sort)
+	if diags.HasError() {
+		return diags
+	}
+	return nil
+}
+
+func refreshObjectRouterBfdNeighbor(d *schema.ResourceData, o *models.RouterBfdNeighbor, sv string, sort bool) diag.Diagnostics {
+	var err error
+
+	if o.Interface != nil {
+		v := *o.Interface
+
+		if err = d.Set("interface", v); err != nil {
+			return diag.Errorf("error reading interface: %v", err)
+		}
+	}
+
+	if o.Ip != nil {
+		v := *o.Ip
+
+		if err = d.Set("ip", v); err != nil {
+			return diag.Errorf("error reading ip: %v", err)
+		}
+	}
+
+	return nil
+}
+
+func getObjectRouterBfdNeighbor(d *schema.ResourceData, sv string) (*models.RouterBfdNeighbor, diag.Diagnostics) {
+	obj := models.RouterBfdNeighbor{}
+	diags := diag.Diagnostics{}
+
+	if v1, ok := d.GetOk("interface"); ok {
+		if v2, ok := v1.(string); ok {
+			if !utils.CheckVer(sv, "", "") {
+				e := utils.AttributeVersionWarning("interface", sv)
+				diags = append(diags, e)
+			}
+			obj.Interface = &v2
+		}
+	}
+	if v1, ok := d.GetOk("ip"); ok {
+		if v2, ok := v1.(string); ok {
+			if !utils.CheckVer(sv, "", "") {
+				e := utils.AttributeVersionWarning("ip", sv)
+				diags = append(diags, e)
+			}
+			obj.Ip = &v2
+		}
+	}
+	return &obj, diags
+}